chore: restructure 'core' module
