+++ ums/api/middlewares/filter_sort.py
@@ -1,44 +1,8 @@
-from enum import Enum
-from typing import Any, Callable, Literal, Type, TypeAlias, get_args
+from typing import Callable, Type, TypeAlias, get_args
 
 from fastapi import Query
-from pydantic import BaseModel
 
-# Filtering
-
-
-class BaseFilterParams(BaseModel):
-    """
-    Base class for filter parameters (this is the parameters container to be
-    used in the query).
-
-    When subclassing, define the filter parameters as fields.
-    """
-
-    def get_filters(self) -> dict[str, Any]:
-        return self.model_dump(exclude_none=True, exclude_unset=True)
-
-
-# Sorting
-
-SortOrder = Literal["asc", "desc"]
-
-
-class SortParams(BaseModel):
-    """
-    Container class for the sorting parameters. To be populated with the
-    `sort_by` and `sort_order` queries.
-    """
-
-    sort_by: str
-    sort_order: SortOrder
-
-
-class SortOptions(str, Enum):
-    """When subclassing, define the available sorting options."""
-
-    ...
-
+from ums.core.utils.filter_sort import SortOptions, SortParams, SortOrder
 
 def parse_sorting(options: Type[SortOptions]) -> Callable:
     OptionsAlias: TypeAlias = options  # type: ignore[valid-type]
