+++ ums/domain/group/writer.py
@@ -1,4 +1,4 @@
-from ums.data_access.writer import GenericWriter
+from ums.core.data_access.writer import GenericWriter
 from ums.domain.entities import Group
