+++ ums/domain/permissions/schemas.py
@@ -1,6 +1,6 @@
 import uuid
 
-from ums.data_access.interfaces import Schema
+from ums.core.data_access.interfaces import Schema
 
 
 class PermissionsCreate(Schema):
