+++ ums/domain/user/validator.py
@@ -1,6 +1,6 @@
 from sqlmodel import column, select
 
-from ums.db.async_session import AsyncSessionStream
+from ums.core.db.async_session import AsyncSessionStream
 from ums.domain import exceptions
 from ums.domain.entities import Group, Role, User
 from ums.domain.user.schemas import UserCreate, UserUpdate
